Introduce a Makefile at the top-level
So far each directory contained a Makefile and at the top-level a bunch
of shell scripts were calling them in a specific way, which not only
involves a lot of redundant stuff but also doesn't resolve dependencies
accordingly (eg. if "install_vsock.sh" is run prior to "doitall" it will
fail).

Now we have a top-level Makefile which just handles installation and
runs the Makefiles in each subdirectory to build. No shell scripts are
involved here anymore and a "make clean" will also recurse accordingly
when done at the top-level.

The deduplication is done via a common.mk in the top-level which is for
building the object files and liking the the target and only a very
minimal set of variables is needed for the Makefiles in the
subdirectories.

Note that I also removed the definition of CC, so that the corresponding
default compiler is used for the current build environment.

Also there is now a PREFIX variable which can be used to install to a
specific directory instead of /usr.

Apart from these changes, there still is a lot to improve and putting
everything in one single Makefile would make it less nasty to do that,
especially when it comes to target propagation. However I kept it like
that as I don't know what the original author's workflow was or is...
maybe they prefer to just type "make" when in the cli or srv directory.

Signed-off-by: aszlig <aszlig@nix.build>